Add django style permissions.

+++ app/src/main/java/com/datavite/distrivite/data/sync/services/BillingSyncServiceImpl.kt
@@ -10,6 +10,7 @@ import com.datavite.distrivite.data.mapper.BillingMapper
 import com.datavite.distrivite.data.remote.datasource.BillingRemoteDataSource
 import com.datavite.distrivite.data.remote.model.RemoteBilling
 import com.datavite.distrivite.data.sync.EntityType
+import com.datavite.distrivite.data.sync.OperationScope
 import com.datavite.distrivite.data.sync.OperationType
 import com.datavite.distrivite.data.sync.SyncConfig
 import kotlinx.coroutines.delay
@@ -229,12 +230,21 @@ class BillingSyncServiceImpl @Inject constructor(
             }
 
             // Remove operation after success
-            pendingOperationDao.deleteByKeys(
-                entityType = currentOperation.entityType,
-                entityId = currentOperation.entityId,
-                operationType = currentOperation.operationType,
-                orgId = currentOperation.orgId
-            )
+            when(currentOperation.operationScope){
+                OperationScope.STATE -> {
+                    pendingOperationDao.deleteByKeys(
+                        entityType = currentOperation.entityType,
+                        entityId = currentOperation.entityId,
+                        operationType = currentOperation.operationType,
+                        orgId = currentOperation.orgId
+                    )
+                }
+                OperationScope.EVENT -> {
+                    pendingOperationDao.deleteById(currentOperation.id)
+                }
+            }
+
+            Log.i("BillingSyncLiedji", "Liedji Successfully synced operation ${currentOperation.operationType} ${currentOperation.entityType} ${currentOperation.entityId}")
 
             // Update sync status depending on remaining operations
             val newStatus = if ((totalPending - 1) > 0) SyncStatus.PENDING else SyncStatus.SYNCED
